feat: add relationship, autorization and authentication

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './task-status.enum';
@@ -12,35 +13,39 @@ export class TasksService {
     @InjectRepository(TaskRepository) private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<TaskEntity[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(    
+    filterDto: GetTasksFilterDto, 
+    user:  User
+    ): Promise<TaskEntity[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: string): Promise<TaskEntity> {
-                                                       const task = await this.taskRepository.findOne(
-                                                         id,
-                                                       );
+  async getTaskById(id: string, user:User): Promise<TaskEntity> {
+    console.log(user.id)
+    const task = await this.taskRepository.findOne({where: { id, userId: user.id }});
 
-                                                       if (!task)
-                                                         throw new NotFoundException();
+    if (!task) throw new NotFoundException();
 
-                                                       return task;
-                                                     }
-
-  async createTask(createTaskDto: CreateTaskDto): Promise<TaskEntity> {
-    return this.taskRepository.createTask(createTaskDto);
+    return task;
   }
 
-  async updateTask(id: string, status: TaskStatus): Promise<TaskEntity> {
-    const task = await this.getTaskById(id);
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<TaskEntity> {
+    return this.taskRepository.createTask(createTaskDto, user);
+  }
+
+  async updateTask(id: string, status: TaskStatus, user:User): Promise<TaskEntity> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
 
     return task;
   }
 
-  async deleteTask(id: string): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: string, user:User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) throw new NotFoundException();
